libtcmu: fix inquiry buffer length calculations
The tcmu device config string is added as an entry in the device
identification page, which has a corresponding one-byte length stored
and later retrieved (for length calculations) from a char ptr.

If device config string length exceeds 127 characters (char maximum)
then subsequent length calculations are done using a wrapped value.

Bug: https://github.com/open-iscsi/tcmu-runner/issues/76

+++ api.c
@@ -477,7 +477,7 @@ int tcmu_emulate_evpd_inquiry(
 		len = snprintf(&ptr[12], sizeof(data) - 16, "%s", wwn);
 
 		ptr[3] = 8 + len + 1;
-		used += ptr[3] + 4;
+		used += (uint8_t)ptr[3] + 4;
 		ptr += used;
 
 		/* 2/3: NAA binary */
@@ -528,7 +528,7 @@ int tcmu_emulate_evpd_inquiry(
 		len = snprintf(&ptr[4], sizeof(data) - used - 4, "%s", dev->cfgstring);
 		ptr[3] = len + 1;
 
-		used += ptr[3] + 4;
+		used += (uint8_t)ptr[3] + 4;
 
 		/* Done with descriptor list */
